refactor(WalletNetwork): move methods out of MyWallet and add tests
Also use promises.
