+++ src/wallet.js
@@ -789,117 +789,6 @@ MyWallet.makePairingCode = function(success, error) {
   }
 };
 
-/**
- * Fetch information on wallet identfier with resend code set to true
- * @param {string} user_guid User GUID.
- * @param {function()} success Success callback function.
- * @param {function()} error Error callback function.
- */
-// used in the frontend
-MyWallet.resendTwoFactorSms = function(user_guid, success, error) {
-
-  var data = {
-    format : 'json',
-    resend_code : true,
-    ct : (new Date()).getTime(),
-    api_code : API.API_CODE
-  }
-  var s = function(obj) { success(); }
-  var e = function(e) {
-    if(e.responseJSON && e.responseJSON.initial_error) {
-      error(e.responseJSON.initial_error);
-    } else {
-      error();
-    }
-  }
-  API.request("GET", 'wallet/'+user_guid, data, true, false).then(s).catch(e);
-};
-
-/**
- * Trigger an email with the users wallet guid(s)
- * @param {string} user_email Registered mail address.
- * @param {string} captcha Spam protection
- * @param {function()} success Success callback function.
- * @param {function()} error Error callback function.
- */
-// used in the frontend
-MyWallet.recoverGuid = function(user_email, captcha) {
-  var defer = Q.defer();
-
-  var data = {
-    method: 'recover-wallet',
-    email : user_email,
-    captcha: captcha,
-    ct : (new Date()).getTime(),
-    api_code : API.API_CODE
-  }
-  var s = function(obj) {
-    if(obj.success) {
-      defer.resolve(obj.message);
-    } else {
-      defer.reject(obj.message);
-    }
-  }
-  var e = function(e) {
-    if(e.responseJSON && e.responseJSON.initial_error) {
-      defer.reject(e.responseJSON.initial_error);
-    } else {
-      defer.reject();
-    }
-  }
-  API.request("POST", 'wallet', data, true).then(s).catch(e);
-
-  return defer.promise;
-};
-
-/**
- * Trigger the 2FA reset process
- * @param {string} user_guid User GUID.
- * @param {string} user_email Registered email address.
- * @param {string} user_new_email Optional new email address.
- * @param {string} secret
- * @param {string} message
- * @param {string} captcha Spam protection
- * @param {function()} success Success callback function.
- * @param {function()} error Error callback function.
- */
-// used in the frontend
-MyWallet.requestTwoFactorReset = function(
-  user_guid,
-  user_email,
-  user_new_email,
-  secret,
-  message,
-  captcha) {
-
-  var defer = Q.defer();
-
-  var data = {
-    method: 'reset-two-factor-form',
-    guid: user_guid,
-    email: user_email,
-    contact_email: user_new_email,
-    secret_phrase: secret,
-    message: message,
-    kaptcha: captcha,
-    ct : (new Date()).getTime(),
-    api_code : API.API_CODE
-  }
-  var s = function(obj) {
-    if(obj.success) {
-      defer.resolve(obj.message);
-    } else {
-      defer.reject(obj.message);
-    }
-  }
-  var e = function(e) {
-    defer.reject(e);
-  }
-  API.request("POST", 'wallet', data, true).then(s).catch(e);
-
-  return defer.promise;
-};
-
 ////////////////////////////////////////////////////////////////////////////////
 MyWallet.login = function ( user_guid
                           , shared_key
